backend/requirements.txt
Flask>=2.0
pymongo>=4.0
APScheduler>=3.8
python-dotenv>=0.21
requests>=2.28
PyJWT>=2.6
python-dateutil>=2.8